* string の文字位置を指定する変数の型を unsigned int から size_t に変更
--HG--
branch : implement_vm

+++ encoding.c
@@ -137,7 +137,7 @@ scm_str_itr_next(const ScmStrItr *iter)
 
 static ScmStrItr
 scm_enc_index2itr_fixed_width(void *str, size_t size,
-                              unsigned int idx, size_t width,
+                              size_t idx, size_t width,
                               int (*char_width)(const void *str, size_t len))
 {
   ScmStrItr iter;
@@ -160,11 +160,11 @@ scm_enc_index2itr_fixed_width(void *str, size_t size,
 
 static ScmStrItr
 scm_enc_index2itr_variable_width(void *str, size_t size,
-                                 unsigned int idx,
+                                 size_t idx,
                                  int (*char_width)(const void *str, size_t len))
 {
   ScmStrItr iter;
-  unsigned int i;
+  size_t i;
 
   iter = scm_str_itr_begin(str, size, char_width);
   if (SCM_STR_ITR_IS_ERR(&iter)) return iter;
@@ -200,7 +200,7 @@ scm_enc_char_width_ascii(const void *str, size_t len)
 }
 
 ScmStrItr
-scm_enc_index2itr_ascii(void *str, size_t size, unsigned int idx)
+scm_enc_index2itr_ascii(void *str, size_t size, size_t idx)
 {
   return scm_enc_index2itr_fixed_width(str, size, idx, sizeof(scm_char_ascii_t),
                                        scm_enc_char_width_ascii);
@@ -237,7 +237,7 @@ scm_enc_char_width_bin(const void *str, size_t len)
 }
 
 ScmStrItr
-scm_enc_index2itr_bin(void *str, size_t size, unsigned int idx)
+scm_enc_index2itr_bin(void *str, size_t size, size_t idx)
 {
   return scm_enc_index2itr_fixed_width(str, size, idx, sizeof(scm_char_bin_t),
                                        scm_enc_char_width_ascii);
@@ -312,7 +312,7 @@ scm_enc_char_width_utf8(const void *str, size_t len)
 }
 
 ScmStrItr
-scm_enc_index2itr_utf8(void *str, size_t size, unsigned int idx)
+scm_enc_index2itr_utf8(void *str, size_t size, size_t idx)
 {
   return scm_enc_index2itr_variable_width(str, size, idx,
                                           scm_enc_char_width_utf8);
@@ -342,7 +342,7 @@ scm_enc_char_width_ucs4(const void *str, size_t len)
 }
 
 ScmStrItr
-scm_enc_index2itr_ucs4(void *str, size_t size, unsigned int idx)
+scm_enc_index2itr_ucs4(void *str, size_t size, size_t idx)
 {
 
   return scm_enc_index2itr_fixed_width(str, size, idx, sizeof(scm_char_ucs4_t),
@@ -449,7 +449,7 @@ scm_enc_char_width_eucjp(const void *str, size_t len)
 }
 
 ScmStrItr
-scm_enc_index2itr_eucjp(void *str, size_t size, unsigned int idx)
+scm_enc_index2itr_eucjp(void *str, size_t size, size_t idx)
 {
   return scm_enc_index2itr_variable_width(str, size, idx,
                                           scm_enc_char_width_eucjp);
@@ -495,7 +495,7 @@ scm_enc_char_width_sjis(const void *str, size_t len)
 }
 
 ScmStrItr
-scm_enc_index2itr_sjis(void *str, size_t size, unsigned int idx)
+scm_enc_index2itr_sjis(void *str, size_t size, size_t idx)
 {
   return scm_enc_index2itr_variable_width(str, size, idx,
                                           scm_enc_char_width_sjis);
